+++ src/features/pacientes/data.js
@@ -1,0 +1,32 @@
+export const columns = [
+  { key: "paciente", label: "Paciente" },
+  { key: "citas", label: "Citas" },
+  { key: "estado", label: "Estado" },
+];
+
+export const rows = [
+  {
+    key: "1",
+    name: "Mario Gomez",
+    date: "Seleccione una cita",
+    status: "En Proceso",
+  },
+  {
+    key: "2",
+    name: "Juan Carlos Osorio",
+    date: "Seleccione una cita",
+    status: "Finalizada",
+  },
+  {
+    key: "3",
+    name: "Mario Gomez",
+    date: "Seleccione una cita",
+    status: "Reprogramada",
+  },
+  {
+    key: "4",
+    name: "Juan Carlos Osorio",
+    date: "Seleccione una cita",
+    status: "Finalizada",
+  },
+];
